+++ label_ptrs.hpp
@@ -2,4 +2,4 @@
 
 #include <unordered_map>
 
-using LabelPtrs = std::unordered_map<int32_t, int32_t>;
+using LabelPtrs = std::unordered_map<int, intptr_t>;
